# Optimize handling of purely-LTR paragraphs