This aims to boost performance for paragraphs that do not contain any strongly-RTL characters.

Typical results of `cargo bench -F bench_it --bench udhr` with current master:

    test bench_1_bidi_info_new_for_ltr_texts  ... bench:     891,763 ns/iter (+/- 43,823)
    test bench_2_bidi_info_new_for_bidi_texts ... bench:   1,002,740 ns/iter (+/- 173,182)
    test bench_3_reorder_line_for_ltr_texts   ... bench:     155,455 ns/iter (+/- 17,878)
    test bench_4_reorder_line_for_bidi_texts  ... bench:     330,877 ns/iter (+/- 39,449)

And the same test after applying this patch:

    test bench_1_bidi_info_new_for_ltr_texts  ... bench:     875,862 ns/iter (+/- 99,252)
    test bench_2_bidi_info_new_for_bidi_texts ... bench:   1,005,841 ns/iter (+/- 179,152)
    test bench_3_reorder_line_for_ltr_texts   ... bench:         894 ns/iter (+/- 57)
    test bench_4_reorder_line_for_bidi_texts  ... bench:     243,300 ns/iter (+/- 16,616)